Revert "Commited changes to support S3 as data store"
