+++ data-access/nexustiles/config/datastores.ini
@@ -4,17 +4,6 @@ keyspace=nexustiles
 local_datacenter=datacenter1
 protocol_version=3
 
-[s3]
-bucket=nexus-jpl
-region=us-west-2
-
-[dynamo]
-table=nexus-jpl-table
-region=us-west-2
-
 [solr]
 host=localhost:8983
 core=nexustiles
-
-[datastore]
-store=s3
