+++ data-access/nexustiles/dao/CassandraProxy.pyx
@@ -1,7 +1,3 @@
-"""
-Copyright (c) 2017 Jet Propulsion Laboratory,
-California Institute of Technology.  All rights reserved
-"""
 import uuid
 from ConfigParser import NoOptionError
 from multiprocessing.synchronize import Lock
