+++ data-access/nexustiles/dao/SolrProxy.pyx
@@ -1,7 +1,3 @@
-"""
-Copyright (c) 2017 Jet Propulsion Laboratory,
-California Institute of Technology.  All rights reserved
-"""
 import json
 import logging
 import threading
