+++ nexus-ingest/nexus-sink/build.gradle
@@ -1,6 +1,5 @@
 buildscript {
     repositories {
-        mavenCentral()
         if( project.hasProperty('artifactory_contextUrl') ) {
             maven {
                 url "${artifactory_contextUrl}"
@@ -32,7 +31,6 @@ buildscript {
     dependencies {
         classpath "org.jfrog.buildinfo:build-info-extractor-gradle:4+"
         classpath("org.springframework.xd:spring-xd-module-plugin:1.3.1.RELEASE")
-        classpath "io.spring.gradle:dependency-management-plugin:1.0.0.RC2"
     }
 }
 
@@ -109,12 +107,13 @@ ext {
     ]
 }
 
+
+
 apply plugin: 'java'
 apply plugin: 'groovy'
 apply plugin: 'idea'
 apply plugin: 'spring-xd-module'
 apply plugin: 'maven-publish'
-apply plugin: "io.spring.dependency-management"
 apply plugin: 'project-report'
 
 group = 'org.nasa.jpl.nexus.ingest'
@@ -146,12 +145,6 @@ sourceSets {
     }
 }
 
-dependencyManagement {
-    imports {
-        mavenBom 'com.amazonaws:aws-java-sdk-bom:1.10.77+'
-    }
-}
-
 //noinspection GroovyAssignabilityCheck
 dependencies {
     compile("org.springframework.boot:spring-boot-starter-integration")
@@ -164,10 +157,8 @@ dependencies {
     compile "org.apache.solr:solr-solrj:5.3.1"
     compile 'org.codehaus.groovy:groovy'
 
-    compile 'com.amazonaws:aws-java-sdk-s3'
-    compile 'com.amazonaws:aws-java-sdk-dynamodb'
-    testCompile 'io.findify:s3mock_2.12:0.2.3'
 
+    testCompile group: 'io.findify', name: 's3mock_2.12', version: '0.2.3'
     testCompile('org.springframework.boot:spring-boot-starter-test')
     testCompile("org.apache.cassandra:cassandra-all")
     testCompile("org.apache.solr:solr-core:${testversions.solrCoreVersion}"){
@@ -178,10 +169,6 @@ dependencies {
         exclude group: 'org.slf4j', module: 'slf4j-log4j12'
     }
     testCompile group: 'junit', name: 'junit', version: '4.11'
-}
-
-task buildScriptDependencies(type: org.gradle.api.tasks.diagnostics.DependencyReportTask) {
-    configurations = project.buildscript.configurations
 }
 
 task wrapper(type: Wrapper) {
