+++ nexus-ingest/nexus-sink/gradle/wrapper/gradle-wrapper.properties
@@ -1,4 +1,4 @@
-#Wed Jun 28 16:08:35 PDT 2017
+#Thu Aug 11 13:56:54 PDT 2016
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 zipStoreBase=GRADLE_USER_HOME
