--- nexus-ingest/nexus-sink/src/main/java/org/nasa/jpl/nexus/ingest/nexussink/DataStoreException.java
@@ -1,24 +1,0 @@
-/*****************************************************************************
- * Copyright (c) 2017 Jet Propulsion Laboratory,
- * California Institute of Technology.  All rights reserved
- *****************************************************************************/
-package org.nasa.jpl.nexus.ingest.nexussink;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-/**
- * Created by djsilvan on 8/11/17.
- */
-public class DataStoreException extends RuntimeException {
-
-    private Logger log = LoggerFactory.getLogger(NexusService.class);
-
-    public DataStoreException() {
-        log.error("Error: DataStore Exception");
-    }
-
-    public DataStoreException(Exception e) {
-        log.error("Error: " + e.getMessage());
-    }
-}
